+++ examples/companion_radio/MyMesh.cpp
@@ -823,9 +823,9 @@ MyMesh::MyMesh(mesh::Radio &radio, mesh::RNG &rng, mesh::RTCClock &rtc, SimpleMe
   //_prefs.rx_delay_base = 10.0f;  enable once new algo fixed
 #if defined(USE_SX1262) || defined(USE_SX1268)
 #ifdef SX126X_RX_BOOSTED_GAIN
-  _prefs.sx126x_rx_boosted_gain = SX126X_RX_BOOSTED_GAIN;
+  _prefs.rx_boosted_gain = SX126X_RX_BOOSTED_GAIN;
 #else
-  _prefs.sx126x_rx_boosted_gain = 1; // enabled by default
+  _prefs.rx_boosted_gain = 1; // enabled by default
 #endif
 #endif
 }
@@ -894,11 +894,9 @@ void MyMesh::begin(bool has_display) {
 
   radio_set_params(_prefs.freq, _prefs.bw, _prefs.sf, _prefs.cr);
   radio_set_tx_power(_prefs.tx_power_dbm);
-#if defined(USE_SX1262) || defined(USE_SX1268)
-  radio_set_rx_boosted_gain_mode(_prefs.sx126x_rx_boosted_gain);
-  MESH_DEBUG_PRINTLN("SX126x RX Boosted Gain Mode: %s",
-                     radio_get_rx_boosted_gain_mode() ? "Enabled" : "Disabled");
-#endif
+  radio_driver.setRxBoostedGainMode(_prefs.rx_boosted_gain);
+  MESH_DEBUG_PRINTLN("RX Boosted Gain Mode: %s",
+                     radio_driver.getRxBoostedGainMode() ? "Enabled" : "Disabled");
 }
 
 const char *MyMesh::getNodeName() {
@@ -1728,7 +1726,7 @@ void MyMesh::handleCmdFrame(size_t len) {
       out_frame[i++] = STATS_TYPE_CORE;
       uint16_t battery_mv = board.getBattMilliVolts();
       uint32_t uptime_secs = _ms->getMillis() / 1000;
-      uint8_t queue_len = (uint8_t)_mgr->getOutboundCount(0xFFFFFFFF);
+      uint8_t queue_len = (uint8_t)_mgr->getOutboundTotal();
       memcpy(&out_frame[i], &battery_mv, 2); i += 2;
       memcpy(&out_frame[i], &uptime_secs, 4); i += 4;
       memcpy(&out_frame[i], &_err_flags, 2); i += 2;